+++ ConsoleGameTests/Mock_GameInfoProvider.h
@@ -10,8 +10,8 @@ public:
    MOCK_METHOD( ConsoleGame::GameState, GetGameState, ( ), ( const, override ) );
    MOCK_METHOD( ConsoleGame::Direction, GetPlayerDirection, ( ), ( const, override ) );
    MOCK_METHOD( bool, IsPlayerMoving, ( ), ( const, override ) );
-   MOCK_METHOD( int, GetArenaWidth, ( ), ( const, override ) );
-   MOCK_METHOD( int, GetArenaHeight, ( ), ( const, override ) );
-   MOCK_METHOD( int, GetArenaPlayerXPosition, ( ), ( const, override ) );
-   MOCK_METHOD( int, GetArenaPlayerYPosition, ( ), ( const, override ) );
+   MOCK_METHOD( double, GetArenaWidth, ( ), ( const, override ) );
+   MOCK_METHOD( double, GetArenaHeight, ( ), ( const, override ) );
+   MOCK_METHOD( double, GetArenaPlayerXPosition, ( ), ( const, override ) );
+   MOCK_METHOD( double, GetArenaPlayerYPosition, ( ), ( const, override ) );
 };
